apps/api/pyproject.toml
[project]
name = "forensic_council"
version = "1.4.0"
description = "Multi-Agent Forensic Evidence Analysis System"
# Upper bound: qdrant-client 1.14.1–1.16.1 force numpy>=2.1 on py3.13+, conflicting
# with moviepy 1.x which requires numpy<2.0. Docker image uses Python 3.12.
requires-python = ">=3.11,<3.15"
license = { text = "MIT" }
authors = [
    { name = "Forensic Council Team" }
]

dependencies = [
    # Core
    "redis>=5.3",                          # async support built-in via redis.asyncio (extra removed in 5.x)
    # qdrant==1.16.2 is the last pre-1.17 release that only requires numpy>=2.3 on
    # Python 3.14+ (not 3.13+). Combined with requires-python<3.13 this is safe with numpy<2.0.
    # [async] extra removed in 1.10+ — async support (httpx) is built-in.
    "qdrant-client==1.16.2",
    "asyncpg>=0.30",
    "pydantic>=2.11",
    "pydantic-settings>=2.9",
    "cryptography>=44",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "httpx>=0.28",
    "groq>=0.28",                      # Groq LLM API client (Llama 3.3 70B)
    "google-generativeai>=0.8",        # Google Gemini vision API (Agents 1, 3, 5 deep analysis)
    # API Server
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.20",
    "websockets>=15.0",
    "python-jose[cryptography]>=3.5",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.1,<5.0",
    # Image processing tools
    "Pillow>=11.2",
    "numpy>=1.26,<3.0",  # relaxed upper bound — scipy 1.17+ requires numpy>=2.0
    "scipy>=1.15",
    "piexif>=1.1",
    "imagehash>=4.3",
    "pytesseract>=0.3.13",
    "opencv-contrib-python>=4.9",    # SIFT copy-move detection
    "scikit-image>=0.25",             # Shadow/lighting Hough transform
    "stegano>=0.11",                  # LSB steganography decode
    "hachoir>=3.2",                   # Binary container metadata parser
    # Metadata tools
    "timezonefinder>=6.2",
    "pyexiftool>=0.5",
    "geopy>=2.4",
    "astral>=3.2",                    # Astronomical validation
    # Audio processing tools
    "librosa>=0.10",
    "soundfile>=0.12",
    "praat-parselmouth>=0.4",         # Gold-standard prosody (F0, jitter, shimmer)
    "moviepy>=1.0,<2.0",             # AV sync via video+audio correlation
    # ML tools (lightweight)
    "scikit-learn>=1.7",
    # File type detection — python-magic-bin bundles libmagic for Windows
    "python-magic>=0.4.27;sys_platform!='win32'",
    "python-magic-bin>=0.4.14;sys_platform=='win32'",
    # OCR & document extraction (three-tier pipeline: PyMuPDF -> EasyOCR -> Tesseract)
    "PyMuPDF>=1.25",                  # Lossless PDF text/image/metadata extraction (fitz)
    "easyocr>=1.7.2",                   # Neural OCR: handles curved text, low-res, 80+ langs
    # AV container profiling
    "pymediainfo>=6.1.0",               # Deep container analysis: codec, fps mode, writing app
]

[project.optional-dependencies]
ml = [
    # ── PyTorch (CPU-only) ──────────────────────────────────────────────────
    # Sources below route these to download.pytorch.org/whl/cpu so only
    # CPU wheels are resolved. This avoids ~5 GB of nvidia/triton/cuda CUDA
    # packages that default PyPI torch pulls in on Linux (torch+cu128).
    "torch>=2.7",
    "torchvision>=0.22",
    "torchaudio>=2.7",
    # ── Vision / Object detection ───────────────────────────────────────────
    "ultralytics>=8.3",
    "open-clip-torch>=2.30",
    # ── NLP / Vision transformers ───────────────────────────────────────────
    "transformers>=4.52",
    "accelerate>=1.7",
    # ── Audio / Speech Forensics ────────────────────────────────────────────
    "speechbrain>=1.0",
    # NOTE: deepface intentionally excluded — it hard-requires tensorflow (1.8 GB
    # CUDA) + opencv-python (conflicts with opencv-contrib-python) + Flask/gunicorn
    # (unused web servers). Face-swap detection in video_tools.py has a graceful
    # fallback when deepface is absent.
]
dev = [
    "pytest>=8.4",
    "pytest-asyncio>=1.0",
    "pytest-cov>=6.1",
    "pytest-mock>=3.14",
    "pyyaml>=6.0",
    "ruff>=0.9",
    "pyright==1.1.400",
    "httpx>=0.28",
    "faker>=26.0",
    "factory-boy>=3.3",
]
observability = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-redis>=0.41b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0",
]
security = [
    "fastapi-csrf-protect>=0.3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["."]
addopts = "-v --tb=short --cov-fail-under=75"
markers = [
    "regression: marks tests as regression tests (deselect with -m 'not regression')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "requires_docker: requires a running Docker stack",
    "requires_network: requires internet access",
    "slow: slow tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["core", "agents", "tools", "orchestration"]
branch = true
omit = [
    "tools/ml_tools/*.py",
    "*/worker.py",
    "*/model_cache.py",
]

[tool.coverage.report]
fail_under = 75
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "\\.\\.\\.",
]

[tool.hatch.build.targets.wheel]
packages = ["core", "agents", "tools", "orchestration", "reports", "api"]

# ── Ruff configuration ─────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "PT"]
ignore = ["E501", "S101"]

# ── Pyright configuration ──────────────────────────────────────────────────
[tool.pyright]
include = ["core", "api", "agents", "tools", "orchestration"]
exclude = ["**/__pycache__", "**/.pytest_cache"]
typeCheckingMode = "basic"
strict = ["core/auth.py", "core/signing.py", "core/circuit_breaker.py"]  # Critical security modules

# ── isort configuration ────────────────────────────────────────────────────
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

# ── uv source overrides ────────────────────────────────────────────────────
# Route torch/torchvision/torchaudio to the official CPU-only PyTorch index.
# This prevents uv from resolving the default GPU wheels from PyPI which
# include ~5 GB of CUDA runtime packages (nvidia-*, triton, cuda-*).
[tool.uv.sources]
torch      = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
torchaudio  = { index = "pytorch-cpu" }

[[tool.uv.index]]
name     = "pytorch-cpu"
url      = "https://download.pytorch.org/whl/cpu"
explicit = true   # only used for packages that explicitly reference this index
